fix(tests): corrige une collision silencieuse de stubs entre Salesforce et SalesforceSync
SalesforcePetitionBulkCsvTest.php et SyncSignaturesToSalesforceTest.php
déclaraient chacun leur propre copie de get_local_user()/update_signature_status().
PHP ne garde que la première déclaration chargée dans un process : si les deux
testsuites se chargent ensemble, la seconde copie (bien que protégée par
function_exists()) se retrouve à utiliser silencieusement les globales de
l'autre fichier au lieu des siennes - pas de fatal, juste un résultat faux
(confirmé via `phpunit --testsuite Salesforce,SalesforceSync`).

Les deux stubs sont maintenant factorisés dans tests/support/local-user-stubs.php,
require_once par les deux fichiers, pour garantir une définition unique quel
que soit l'ordre de chargement.
